# internal/api/pdf.go
// filename: pdf.go

package api

import (
	"context"
	"database/sql"
	"encoding/base64"
	"encoding/json"
	"fmt"
	"net/url"
	"strings"
	"time"

	"github.com/gofiber/fiber/v2"
	// "github.com/illegalcall/task-master/internal/models"
)

// ----------------------------------------------------------------------------
// Mock or minimal definitions for demonstration
// (Replace these with your actual project structs/config/db/etc.)
// ----------------------------------------------------------------------------

// JSONRaw is a type alias for json.RawMessage (makes Swagger happy)
type JSONRaw json.RawMessage

// ParseDocumentPayload is the incoming JSON payload for PDF parse jobs
type ParseDocumentPayload struct {
    PDFSource      string  `json:"pdf_source"`
    ExpectedSchema JSONRaw `json:"expected_schema"`
    WebhookURL     string  `json:"webhook_url,omitempty"`
}

// Job represents a generic job in this example
type Job struct {
    ID     int
    Name   string
    Status string
    Type   string
}

// Some constants for job status/types
const (
    StatusPending   = "pending"
    StatusCompleted = "completed"
    JobTypePDFParse = "pdf_parse"
)

// Minimal placeholders to make the sample compile:
type Database struct {
    DB    *sql.DB
    Redis RedisClient
}
type RedisClient interface {
    Set(ctx context.Context, key string, value interface{}, expiration time.Duration) *DummyCmd
}
type DummyCmd struct{}

// Simulate no-op
func (d *DummyCmd) Err() error { return nil }

type Storage struct{}

// StoreFromURL simulates storing a PDF from remote URL
func (s *Storage) StoreFromURL(ctx context.Context, url string) (string, error) {
    return "/mock/path/for/url.pdf", nil
}

// StoreFromBytes simulates storing a PDF from []byte
func (s *Storage) StoreFromBytes(ctx context.Context, data []byte) (string, error) {
    return "/mock/path/from/bytes.pdf", nil
}

// Delete simulates deleting a file
func (s *Storage) Delete(ctx context.Context, path string) error {
    return nil
}

type Config struct {
    Storage struct {
        TTL time.Duration
    }
    Kafka struct {
        Topic string
    }
}

// ----------------------------------------------------------------------------
// Actual handler code
// ----------------------------------------------------------------------------

const maxPDFSize = 10 * 1024 * 1024 // 10MB

// handlePDFParseJob handles PDF parsing job creation
// @Summary Create PDF parsing job
// @Description Creates a new job for parsing a PDF document
// @Tags pdf
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param payload body ParseDocumentPayload true "PDF parsing configuration"
// @Success 200 {object} map[string]interface{} "Job created successfully"
// @Failure 400 {object} map[string]interface{} "Invalid request"
// @Failure 500 {object} map[string]interface{} "Server error"
// @Router /jobs/parse-document [post]
func (s *Server) handlePDFParseJob(c *fiber.Ctx) error {
    ctx := c.Context()

    // Parse and validate the request payload
    var payload ParseDocumentPayload
    if err := c.BodyParser(&payload); err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(map[string]interface{}{
            "error": "Invalid request body",
        })
    }

    // Validate required fields
    if err := validatePDFParsePayload(&payload); err != nil {
        return c.Status(fiber.StatusBadRequest).JSON(map[string]interface{}{
            "error": err.Error(),
        })
    }

    // Store the PDF file
    var pdfPath string
    var err error

    if strings.HasPrefix(payload.PDFSource, "http://") || strings.HasPrefix(payload.PDFSource, "https://") {
        // Handle URL source
        pdfPath, err = s.storage.StoreFromURL(ctx, payload.PDFSource)
    } else {
        // Handle base64 source
        pdfData, decodeErr := base64.StdEncoding.DecodeString(payload.PDFSource)
        if decodeErr != nil {
            return c.Status(fiber.StatusBadRequest).JSON(map[string]interface{}{
                "error": "Invalid base64-encoded PDF data",
            })
        }
        pdfPath, err = s.storage.StoreFromBytes(ctx, pdfData)
    }

    if err != nil {
        return c.Status(fiber.StatusInternalServerError).JSON(map[string]interface{}{
            "error": fmt.Sprintf("Failed to store PDF: %v", err),
        })
    }

    // Create a new job
    job := Job{
        Name:   "PDF Parse Job",
        Status: StatusPending,
        Type:   JobTypePDFParse,
    }

    // Insert job into a mock DB
    // Replace with your real DB insertion code
    job.ID = 123 // mock ID from DB

    // Store job payload in Redis (mock)
    jobPayload := struct {
        ParseDocumentPayload
        PDFPath string `json:"pdf_path"`
    }{
        ParseDocumentPayload: payload,
        PDFPath:              pdfPath,
    }

    payloadBytes, _ := json.Marshal(jobPayload)
    redisKey := fmt.Sprintf("job:%d:payload", job.ID)
    if err := s.db.Redis.Set(ctx, redisKey, payloadBytes, s.cfg.Storage.TTL).Err(); err != nil {
        // Clean up the stored file on error
        _ = s.storage.Delete(ctx, pdfPath)
        return c.Status(fiber.StatusInternalServerError).JSON(map[string]interface{}{
            "error": "Failed to store job payload",
        })
    }

    // Set initial status in Redis (mock)
    statusKey := fmt.Sprintf("job:%d", job.ID)
    if err := s.db.Redis.Set(ctx, statusKey, StatusPending, 0).Err(); err != nil {
        _ = s.storage.Delete(ctx, pdfPath)
        return c.Status(fiber.StatusInternalServerError).JSON(map[string]interface{}{
            "error": "Failed to set job status",
        })
    }

    // Send to Kafka (mock)
    // jobBytes, _ := json.Marshal(job)
    // msg := &sarama.ProducerMessage{
    //     Topic: s.cfg.Kafka.Topic,
    //     Value: sarama.StringEncoder(jobBytes),
    // }
    // Here you'd do: s.producer.SendMessage(msg)
    // We'll just pretend it succeeded.

    // Schedule file cleanup after TTL (mock)
    go func(filePath string, ttl time.Duration) {
        time.Sleep(ttl)
        _ = s.storage.Delete(context.Background(), filePath)
    }(pdfPath, s.cfg.Storage.TTL)

    return c.JSON(map[string]interface{}{
        "job_id": job.ID,
        "status": job.Status,
    })
}

// validatePDFParsePayload validates the PDF parse job payload
func validatePDFParsePayload(payload *ParseDocumentPayload) error {
    // Validate PDF source
    if payload.PDFSource == "" {
        return fmt.Errorf("pdf_source is required")
    }

    // Determine and validate source type
    if strings.HasPrefix(payload.PDFSource, "http://") || strings.HasPrefix(payload.PDFSource, "https://") {
        // Validate URL
        if _, err := url.ParseRequestURI(payload.PDFSource); err != nil {
            return fmt.Errorf("invalid PDF URL")
        }
    } else {
        // Validate base64
        decoded, err := base64.StdEncoding.DecodeString(payload.PDFSource)
        if err != nil {
            return fmt.Errorf("invalid base64-encoded PDF data")
        }

        // Check file size
        if len(decoded) > maxPDFSize {
            return fmt.Errorf("PDF size exceeds maximum allowed size of 10MB")
        }

        // Validate PDF magic number
        if len(decoded) < 4 || string(decoded[:4]) != "%PDF" {
            return fmt.Errorf("invalid PDF format")
        }
    }

    // Validate expected schema
    if len(payload.ExpectedSchema) == 0 {
        return fmt.Errorf("expected_schema is required")
    }

    // Validate that expected_schema is valid JSON
    var js json.RawMessage
    if err := json.Unmarshal(payload.ExpectedSchema, &js); err != nil {
        return fmt.Errorf("invalid JSON schema")
    }

    // Validate webhook URL if provided
    if payload.WebhookURL != "" {
        if _, err := url.ParseRequestURI(payload.WebhookURL); err != nil {
            return fmt.Errorf("invalid webhook URL")
        }
    }

    return nil
}